requirements.txt
Flask==3.0.3
gunicorn==23.0.0
cloudevents==1.11.0
google-cloud-storage==2.16.0



